完善Gallery示例代码
