+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ImageViewer">
         <activity
-            android:name="com.origeek.viewerDemo.MainActivity"
+            android:name="com.origeek.viewerDemo.PreviewerActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ImageViewer.NoActionBar">
@@ -23,15 +23,15 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.origeek.viewerDemo.RemakeActivity"
+            android:name="com.origeek.viewerDemo.GalleryActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ImageViewer.NoActionBar">
-            <!--            <intent-filter>-->
-            <!--                <action android:name="android.intent.action.MAIN" />-->
+<!--            <intent-filter>-->
+<!--                <action android:name="android.intent.action.MAIN" />-->
 
-            <!--                <category android:name="android.intent.category.LAUNCHER" />-->
-            <!--            </intent-filter>-->
+<!--                <category android:name="android.intent.category.LAUNCHER" />-->
+<!--            </intent-filter>-->
         </activity>
         <activity
             android:name="com.origeek.viewerDemo.HugeActivity"
